Add files via upload
This file contains the function for the first two steps in the first stage (Pre-processing and Non-convex Regularisation) of the experiment described in the article. 
The file contains two functions to crop the redundant background (step 1, function crop_image) and to center the image around the probe (step 2, function center_image). The function also contains code to read video files (tested on both avi and gif) read a single frame, perform aforementioned processing and then display the resulting image.